+++ src/main/resources/db/migration/V1__create_entities_tables.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS tb_customers
 (
-    customer_id SERIAL PRIMARY KEY,
+    customer_id BIGSERIAL  PRIMARY KEY,
     name        VARCHAR(255),
     email       VARCHAR(255),
     address     VARCHAR(255)
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS tb_customers
 
 CREATE TABLE IF NOT EXISTS tb_products
 (
-    product_id SERIAL PRIMARY KEY,
+    product_id BIGSERIAL  PRIMARY KEY,
     name       VARCHAR(255),
     quantity   INT,
     version    INT
@@ -16,8 +16,8 @@ CREATE TABLE IF NOT EXISTS tb_products
 
 CREATE TABLE IF NOT EXISTS tb_orders
 (
-    order_id    SERIAL PRIMARY KEY,
-    customer_id INT,
+    order_id    BIGSERIAL  PRIMARY KEY,
+    customer_id BIGINT,
     order_date  TIMESTAMP,
     FOREIGN KEY (customer_id)
         REFERENCES tb_customers (customer_id)
@@ -27,9 +27,9 @@ CREATE TABLE IF NOT EXISTS tb_orders
 
 CREATE TABLE IF NOT EXISTS tb_order_items
 (
-    order_item_id SERIAL PRIMARY KEY,
-    order_id      INT,
-    product_id    INT,
+    order_item_id BIGSERIAL  PRIMARY KEY,
+    order_id      BIGINT,
+    product_id    BIGINT,
     quantity      INT,
     FOREIGN KEY (order_id) REFERENCES tb_orders (order_id),
     FOREIGN KEY (product_id) REFERENCES tb_products (product_id)
